requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
psycopg2-binary==2.9.9
redis==5.1.1
requests==2.32.3
urllib3==2.2.3
websockets==13.1